rename class from mssql to sql
This class is not dedicated to mssql but is used by all sql databases. Rename to reflect this.
