+++ mRemoteNG/Config/Serializers/ConnectionSerializers/Sql/LocalConnectionPropertiesModel.cs
@@ -1,4 +1,4 @@
-﻿namespace mRemoteNG.Config.Serializers.ConnectionSerializers.MsSql
+﻿namespace mRemoteNG.Config.Serializers.ConnectionSerializers.Sql
 {
     public class LocalConnectionPropertiesModel
     {
